+++ CHANGELOG.md
@@ -65,3 +65,10 @@ which block all requests data thread)
 
 * Hotfixes for `0.4.0` version
 * Update `Kotlin Coroutines` -> 0.25.3
+
+## 0.4.2
+
+* Fixes in scheduler plugin
+* Added `/getPublishSchedule` command
+* Added `/disableSchedulePublish` command
+* Now publisher will automatically remove `/publishPost` message